Removes code attempting to simulate fished state
For some of the operation classes to properly parse the response the
operation must be put into the finished state.  This comments out an
attempt at putting the operation in the finished state.  If you attempt
to start the operation it will create and start a NSURLconnection which
you don't want to do.  Currently if you need the operation to be in the
finshed state you need to reopen the operation class and override the
-isFinished method to return true.

e.g.
@interface AFRequestOperation (Spec)
@end

and

@implementation AFRequestOperation (Spec)
-(BOOL)isFinished { return YES; }
@end